Improve the UX for GHA cache restore timeout errors
Previously if any of the cache restore steps failed due to a timeout,
the restore step would only emit a warning, and then the job would
continue onto later steps that will fail if the restored files don't exist.

Now:
* If the cache cannot be restored, the cache restore step itself will
   correctly fail (since we need the restored files).
* The timeout has also been reduced from 10 minutes to 1 minute,
   so the job fails (and can be retried) sooner.

This is a port of the improvements in:
https://github.com/heroku/builder/pull/360

Fixes #61.
GUS-W-13738824.