Fix rust caching and duplicate buildpack compilation (#91)
* Fix rust caching and duplicate buildpack compilation

This is a restructuring of the whole release process to eliminate the issues outlines in #48 (buildpacks compiled multiple times could result in non-determinism) and #49 (rust caching isn't working due to non-unique keys).

Many of the overall steps are still the same but the `detect` job has been replaced by a `compile` job which front loads all the rust and buildpack compilation and cached the compiled buildpack folder for later use. Compilation was previously handled in the `package` job which is no longer needed.

To support this new flow, a `buildpack_output_dir` value had to be added to the generated buildpack matrix. This also allowed some simplification in terms of the assets cached since the OCI image and CNB file can be generated directly from the compiled `buildpack_output_dir` we don't need to store those values in the cache anymore.

* Update .github/workflows/_buildpacks-release.yml

Co-authored-by: Ed Morley <501702+edmorley@users.noreply.github.com>

* Update .github/workflows/_buildpacks-release.yml

Co-authored-by: Ed Morley <501702+edmorley@users.noreply.github.com>

* Update .github/workflows/_buildpacks-release.yml

Co-authored-by: Ed Morley <501702+edmorley@users.noreply.github.com>

* Pull request review comments

* Pull request review comments

---------

Co-authored-by: Ed Morley <501702+edmorley@users.noreply.github.com>