Include the release changelog in the commit message too (#110)
Previously the changelog for a CNB release was only included in the PR
description, and not the commit message body. 

Now the changelog is included in the commit message body too, which
means it's possible to understand what changed when looking at the
commit view, without needing to open up the PR to read there instead.

GUS-W-13912483.