Pin the `libcnb-cargo` version to the same version as `libcnb-package`
The workflows in this repo use libcnb in two ways:
1. As a library (specifically `libcnb-package` and `libcnb-data`
2. As a binary (`cargo libcnb package`)

The version of (1) is controlled by `Cargo.toml` / `Cargo.lock` in this repo.

Previously the version of (2) wasn't pinned at all, and instead the latest version
would be used, meaning that version could get out of sync with that of (1).

Now, the version is manually pinned to the same version as in `Cargo.toml`.

Longer term this should either be replaced by deriving the version in the
workflow, or at least having a CI check/lint that ensures the two are in sync.

However, for the short term we want to just fix CNB releases after the 0.14.0
 libcnb release, until we have time to make the necessary changes in #120.