Fix error message shown when image repository metadata is missing
As of #88 the metadata is now under `image.repository` not `docker.repository`,
however, the error message still referenced the old location.

This caused confusion in when debugging heroku/buildpacks-jvm#60:
https://github.com/heroku/buildpacks-jvm/actions/runs/6626786735/job/18000413851#step:10:10

In addition to fixing the error message, I've renamed the enum variants
and variables to match the new name.

Lastly, I've updated the README to match the more concise release
metadata style now used by all of our buildpacks (the repository URL
is also a real example, to make it clearer that the URL shouldn't include
the `docker://` prefix).

Fixes #168.
GUS-W-14355165.