Allow configuring the PR reviewer for cnb-builder-images PRs (#289)
Currently for CNB releases, the PR opened against `cnb-builder-images`
doesn't have an explicit reviewer set by the automation, which means
it uses that repo's `CODEOWNERS` default of requesting review from
the whole Languages team.

Now, the workflow accepts an optional `reviewers` input, which can
be used to select a specific person (eg the language owner) for review.

This will help reduce review-request-spam to other team members,
and make it easier to see who each PR is waiting on in eg the Slack
review reminders.

See:
https://github.com/peter-evans/create-pull-request?tab=readme-ov-file#action-inputs
https://docs.github.com/en/actions/writing-workflows/workflow-syntax-for-github-actions#onworkflow_dispatchinputs

GUS-W-18011095.