Only unpublish temp tags off Docker Hub if they were created earlier (#307)
* Only unpublish temp tags off Docker Hub if they were created earlier

This also helps work around a current bug/outage Docker Hub are having where deletion attempts fail with an error:

{"message":"operation requires admin permissions","errinfo":{}}

This occurs even if the tags aren't there (so re-running the workflow doesn't work, even though it tries to remove different temporary tags in the re-run).

Assuming this bug remained, it would still cause a workflow failure upon first run, but a re-run (which hits the "already published to Docker Hub" check) would then succeed.

GUS-W-18970431

* clarify 'published_to_' variable name in workflow

it's a bit confusing to read otherwise (published_to_docker would mean it got published to docker earlier in the workflow, not that the tag already existed)